refactor(util): cleanup dir and add debounce fn
