+++ packages/util/src/__specs__/color.spec.js
@@ -1,4 +1,4 @@
-import { transparentize } from '../color.js'
+import { transparentize } from '..'
 
 describe('#transparentize', () => {
   test('happy path', () => {
