+++ packages/util/src/debounce.js
@@ -1,0 +1,12 @@
+export function debounce(delay, fn) {
+  let timerId
+
+  return function(...args) {
+    if (timerId) clearTimeout(timerId)
+
+    timerId = setTimeout(() => {
+      fn(...args)
+      timerId = null
+    }, delay)
+  }
+}
