+++ packages/util/src/index.js
@@ -1,9 +1,10 @@
-export { transparentize } from './color.js'
-export { generatePropTypes } from './prop-defs.js'
-export { whitelistProps } from './props.js'
+export { callAll } from './call-all.js'
 export { capitalize } from './string.js'
-export { useCombinedRefs } from './useCombinedRefs.js'
-export { useToggle } from './useToggle.js'
-export { callAll } from './callAll.js'
-export { createUniversalPortal } from './createUniversalPortal.js'
-export { combineFns } from './combineFns.js'
+export { combineFns } from './combine-fns.js'
+export { createUniversalPortal } from './create-universal-portal.js'
+export { debounce } from './debounce.js'
+export { generateDefaultProps, generatePropTypes } from './prop-defs.js'
+export { transparentize } from './color.js'
+export { useCombinedRefs } from './use-combined-refs.js'
+export { useToggle } from './use-toggle.js'
+export { whitelistProps } from './props.js'
